Bug 1812488 - Do less work for images that we can reuse without revalidation. r=tnikkel
That codepath is hot in some cases like in bug 1812422 and we were doing
a fair bit of extra work like validating the whole security info etc.

If we know that the document involved already loaded the request we
don't need to check all that stuff.

We reorder the checks a bit to keep behavior, validating the security
info first hand, which allows us avoid some complexity (because that
does the CSP check).

Differential Revision: https://phabricator.services.mozilla.com/D167883