Bug 1961140 - Null out ASR for view transition captured contents. r=nical,mstange,botond
We don't want these to scroll relative to their regular ASR or anything
really, as they get captured and removed from the rendering
conceptually.

This is needed to avoid "finite clip" assertions with the previous patch, but
it also fixes cases when scrolling creates bad rendering during live capturing
(e.g., this fixes the test-case in bug 1960839 for me).

Differential Revision: https://phabricator.services.mozilla.com/D245912