Bug 1961181 - Update StyleBench. r=#layout,#perftest-reviewers!
This includes a new media query tests, and two changes that are relevant
to this regression:

 * https://github.com/WebKit/WebKit/commit/b5b511804c67cfbf67f8390bd8564448aee45d6f
 * https://github.com/WebKit/WebKit/pull/44550

This makes sure this performance test measures what's intended to
measure.

Also updates mozilla.patch to account for
https://hg.mozilla.org/mozilla-central/rev/d40e89f84d6f59cdd9ba248feccb3f2bf6d48c74,
which landed without updating it, and updates README_MOZILLA to account
for WebKit's move to git.

Differential Revision: https://phabricator.services.mozilla.com/D246831