Use the non-native theme on windows.
This effectively aligns with Linux, removing the ability to draw
native-looking buttons and progress-bars on Windows.

Note that checkboxes and menus already use non-native drawing, so this
only affects "secondary" UI: the old profile manager, a couple dialogs.

Note also that we already effectively fallback to the non-native theme
dark mode, and that our non-native theme looks reasonably windows-like.

Advantages (apart from not having to maintain all this code) is that the
non-native theme is also faster, and this paves the way towards making
the layout side of theming simpler.

This is to be landed after the soft-freeze of course.